Remove a g_return_val_if_fail() which isn't needed, made gaim_dnsquery_get_port() return an unsigned short and return 0 if query_data is NULL, and fixed a whitespace mistake
--HG--
extra : convert_revision : 0cf86336325db7dcb5ba69d37d700ab7cb957224

+++ libpurple/dnsquery.h
@@ -121,7 +121,7 @@ char *gaim_dnsquery_get_host(GaimDnsQueryData *query_data);
  * @param query_data The DNS query
  * @return The port.
  */
-int gaim_dnsquery_get_port(GaimDnsQueryData *query_data);
+unsigned short gaim_dnsquery_get_port(GaimDnsQueryData *query_data);
 
 /**
  * Initializes the DNS query subsystem.
